==> PD1_Treat/01a_SS2/CellChat/02_CellChat_ForCluster.Rmd <==
---
title: "R Notebook"
output: html_document
---

```{r}
library(CellChat)
library(Seurat) 

options(stringsAsFactors = FALSE)

base_dir = "/Users/jlabelle/"
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Scripts/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/PD1_Treat/SS2/Tumor/03_CellChat/")


analysis_dir<- paste0(working_dir, "/analysis/CellChat/")


cellchatHelper<- paste0(base_dir, script_dir, "CellChat_HelperFunctions.R")
source(cellchatHelper)

setwd(working_dir)

```

```{r}
## args: seurat_dir seurat_obj_name CellAnnot_column_in_seurat analysis_name output_dir file_with_comparisons receiver_cell_type sender_cell_type
## file_with_comparisons should be a text file with 3 columns that follow the structure: comparison_name sender receiver
```

```{r}
args=commandArgs(trailingOnly=TRUE)

seurat_dir<- args[1]
seurat_obj_name<- args[2]
annotColumn<- args[3]
full_analysis_name<- args[4]
output_dir<- args[5]
comparisons_file<- args[6]
receiver_cell_type<- args[7]
sender_cell_type<- args[8]

seurat_dir<- paste0(working_dir, "analysis/")
seurat_obj_name<- "seurat_list_bySample.Rds"
annotColumn<- "immunePrograms_tumorCellTypes"
full_analysis_name<- "ByProgram"
comparisons_file<- "analysis/all_comparisons.txt"
receiver_cell_type<- "AllReceiver"
sender_cell_type<- "AllSender"
```

## Load in list of seurat objects
## 1 object for every sample, with tumor + Tcell data combined
```{r}
seurat_list<- readRDS(paste0(seurat_dir, seurat_obj_name))

## Set column to use as cell annotations
## Can use "CellAnnot" for merged annotations (like OPC-like*) or "OriginalAnnot" for completely un-merged annots
seurat_list<- lapply(seurat_list, function(x){
  tmp<-x
  tmp$CellAnnot<- tmp@meta.data[,annotColumn]
  return(tmp)
})

## Load in desired comparisons
comparisons<- read.table(comparisons_file)
comparison_names<- comparisons$V1
comparisons<- as.list(as.data.frame(t(comparisons[,2:3])))
names(comparisons)<- comparison_names
```

## Settings
1. CellChat settings
```{r}
## CellChat settings
Pvaluethresh<- 1
PCthresh<-0.2
raw.use=TRUE
population.size = FALSE
min.cells = 10
only.pos=TRUE
```

## Initialize settings df
```{r}
settings_df<- data.frame(PvalueThres=Pvaluethresh,
                         PCthresh=PCthresh,
                         raw.use=as.character(raw.use),
                         population.size=as.character(population.size),
                         min.cells=as.integer(min.cells),
                         only.pos=only.pos,
                         comparison_file=comparisons_file)

```


## For each comparison:
1. Set analysis name
2. Set subtypes that contain both cell types
```{r}
comparisons_metadata<- lapply(names(comparisons), function(x){
  comparison<- comparisons[[x]]
  
  ## Make analysis name that will be used for everything downstream for this comparison
  analysisName<- paste0(sender_cell_type, "_", 
                        gsub(" ", "", comparison[1]), ".",
                        receiver_cell_type, "_", 
                        gsub(" ", "", comparison[2]))
  
  ## Create list with analysis name/subtypes/comparison to be used to run CellChat
  comparison_info<- list(analysisName=analysisName,
                         comparison=comparison)
  return(comparison_info)
})
names(comparisons_metadata)<- names(comparisons)

```


## For each comparison:
1. Subset to just samples with at least 10 cells in each cell type of interest
2. Create cellChat object for each of these samples
```{r}
allCellChat<- lapply(comparisons_metadata, function(meta){
  seurat<-seurat_list
  comparison<- meta$comparison
  print(paste0("Processing comparison: ", comparison[1], " vs ", comparison[2]))
        
  ## remove any samples that don't have at least 10 cells for each cell type of interest
  nCellsOfInterest<- lapply(seurat, function(x){
          tbl<- table(x$CellAnnot)
          c(tbl[comparison[1]], tbl[comparison[2]])
  })

  samplePF<- lapply(nCellsOfInterest, function(x){
                  x[comparison[1]]>10 & x[comparison[2]] >10})
  seurat<- seurat[unlist(samplePF) & !(is.na(samplePF))]
  
  all_samples<-names(seurat)
  print(paste0("Samples used: ", paste(all_samples, collapse=", ")))
      
  ## Subset seurat to just cells of interest
  seurat<- lapply(seurat, function(x){
    subset(x, CellAnnot %in% comparison)
  })
        
  ## Create object, preprocess
  all_cellChat<- lapply(all_samples, function(x){
    CreateCellChatObject(mySample = x, 
                         all_merged_seurat_obj = seurat,
                         SpecifyCellAnnot = FALSE, ## already subset to cells of interest
                         Pvaluethresh = Pvaluethresh,
                         PCthresh =PCthresh,
                         only.pos=only.pos) 
  })
  names(all_cellChat)<- all_samples
  
  return(all_cellChat)})

names(allCellChat)<- names(comparisons_metadata)
saveRDS(allCellChat, file=paste0(analysis_dir, "allCellChat.Rds"))
```


## For each cellchat object built for each comparison:
1. Add settings df to each object
2. Run CellChat 
```{r}
allCellChat<- readRDS(paste0(analysis_dir, "allCellChat.Rds"))

all_cellChat_interact<- lapply(names(allCellChat), function(comparisonName){
  all_cellChat<- allCellChat[[comparisonName]]
  comparison<- comparisons_metadata[[comparisonName]]$comparison
  analysisName<-comparisons_metadata[[comparisonName]]$analysisName
  
  message(paste0("Computing interaction network for ", comparison[[1]]," vs ", comparison[[2]]))
  
  ## Compute interaction network
  cellChat<- all_cellChat
    
  ## skip if no cellchat objects for that subtype
  if(class(cellChat) != "list"){next}
  
  cellChat_interact<- lapply(cellChat, function(x){
          InferCellCellInteractions(x, 
                                    raw.use=raw.use, 
                                    population.size = population.size, 
                                    min.cells = min.cells)
  })
  
  ## Save settings to object
  settings_df_ind<- settings_df
  settings_df_ind$cellTypesUsed<- paste(comparison, collapse=".")
  settings_df_ind$analysisName<- analysisName
  settings_df_ind<- as.data.frame(t(settings_df_ind))
  
  
  cellChat_interact[["Settings"]]<- settings_df_ind
  
  return(cellChat_interact)
})

names(all_cellChat_interact)<-names(allCellChat)


saveRDS(all_cellChat_interact, file=paste0(analysis_dir, "/all_cellChat_interact_", 
                                       full_analysis_name, ".Rds"))    
```
